requirements.txt
langchain
langchain-community
streamlit
langchain-openai
tiktoken
python-dotenv
faiss-cpu
